backend/requirements.txt
asgiref==3.7.2
attrs==24.2.0
Automat==24.8.1
certifi==2024.7.4
cffi==1.17.0
chardet==3.0.4
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.0
cssselect==1.2.0
defusedxml==0.7.1
Django==5.0.3
django-3-jet==1.0.8
django-admin-interface==0.29.4
django-ckeditor==6.7.2
django-colorfield==0.11.0
django-cors-headers==4.3.1
django-jazzmin==3.0.1
django-jet==1.0.8
django-js-asset==3.0.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
Faker==24.2.0
filelock==3.15.4
googletrans==4.0.0rc1
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.11.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==24.7.2
itemadapter==0.9.0
itemloaders==1.3.1
jmespath==1.0.1
langdetect==1.0.9
lxml==5.3.0
mysqlclient==2.2.7
packaging==24.1
parsel==1.9.1
pillow==10.3.0
Protego==0.3.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyDispatcher==2.0.7
PyJWT==2.8.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
rfc3986==1.5.0
Scrapy==2.11.2
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
text-unidecode==1.3
tldextract==5.1.2
Twisted==24.7.0
typing_extensions==4.12.2
urllib3==2.2.2
w3lib==2.2.1
whitenoise==6.8.2
zope.interface==7.0.1
